requirements.txt
# Core & Async
asyncio
httpx[http2]
tenacity
loguru
python-dotenv
pyyaml
click
requests

# Media Processing
streamlink
faster-whisper
soundfile
numpy
scipy
scenedetect[opencv]
torch
aiofiles

# Storage
aiobotocore

# TikTok Publishing
# Nota: La librería oficial de TikTok no soporta upload programático.
# Se usará httpx para llamadas directas a la API.

# Dev & Testing
pytest
pytest-asyncio
ruff